LayoutTests/accessibility/custom-elements/multiselectable-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("grid-1").isMultiSelectable is true
PASS accessibilityController.accessibleElementById("grid-2").isMultiSelectable is false
PASS successfullyParsed is true

TEST COMPLETE
